+++ src/hooks/main/useMain.ts
@@ -1,8 +1,7 @@
 import { showToast } from "@src/libs/toast/swal";
 import { useState } from "react";
 import { useGetSeniortojuniorList } from "src/queries/seniortojunior/seniortojunior.quey";
-import { pageParms } from "src/repositories/SeniortojuniorRespository/SeniortojuniorRepository";
-
+import { pageParms } from "src/types/util/pageParms.type";
 
 export const UseMain=()=>{
     const  pageParms={
