+++ src/repositories/SeniortojuniorRespository/SeniortojuniorRepository.ts
@@ -1,9 +1,5 @@
 import { SeniortojuniorResponse, SeniortojuniorTypes } from "src/types/seniortojunior/seniortojunior.type";
-
+import { pageParms } from "src/types/util/pageParms.type";
 export interface SeniortojuniorRepository{
     getSeniortojunior(pageParms:pageParms):Promise<SeniortojuniorResponse>
-}
-export interface pageParms{
-    page:number;
-    size:number;
 }
