+++ src/types/util/pageParms.type.ts
@@ -1,0 +1,4 @@
+export interface pageParms{
+    page:number;
+    size:number;
+}
